Move JSONParsingOptions to parsing package
